Removed unnecessary code. Renamed Base Type filter component
