--- core-web/libs/portlets/dot-content-drive/portlet/src/lib/components/dot-content-drive-toolbar/components/dot-content-drive-baseType-selector/dot-content-drive-baseType-selector.component.scss
@@ -1,7 +1,0 @@
-:host ::ng-deep {
-    width: 15em;
-
-    p-checkbox.p-element {
-        gap: 0.1em;
-    }
-}
