+++ core-web/libs/portlets/dot-content-drive/portlet/src/lib/components/dot-content-drive-toolbar/dot-content-drive-toolbar.component.html
@@ -2,8 +2,7 @@
     <div class="p-toolbar-group-start">
         <dot-content-drive-tree-toggler data-testid="tree-toggler" />
 
-        <!-- While you develop, just replace the div with the intended component. -->
-        <dot-content-drive-base-type-selector />
+        <dot-content-drive-base-type-selector data-testid="base-type-selector" />
         <dot-content-drive-content-type-field data-testid="content-type-field" />
         <dot-content-drive-search-input data-testid="search-input" />
     </div>
